Struct ladspa::Port

pub struct Port {
    pub name: &'static str,
    pub desc: PortDescriptor,
    pub hint: Option<ControlHint>,
    pub default: Option<DefaultValue>,
    pub lower_bound: Option<Data>,
    pub upper_bound: Option<Data>,
}

Represents an input or output to the plugin representing either audio or control data.

Fields

name

The name of the port. For control ports, this will likely be shown by the host in an automatically generated GUI next to the control. For audio ports, it is mostly just for identification in your code but some hosts may display it.

desc

Describes the type of port: audio or control, input or output.

hint

Most useful on control inputs but can be used on any type of port.

default

Most useful on control inputs but can be used on any type of port.

lower_bound

The lower bound of values to accepted by default (the host may ignore this).

upper_bound

The upper bound of values to accepted by default (the host may ignore this).