Trait ladspa::Plugin

pub trait Plugin {
    fn run<'a>(&mut self, sample_count: usize, ports: &[&'a PortConnection<'a>]);

    fn activate(&mut self) { ... }
    fn deactivate(&mut self) { ... }
}

Represents an instance of a plugin which may be exposed as a LADSPA plugin using get_ladspa_descriptor. It is not necessary to implement activate to deactivate.

Required Methods

fn run<'a>(&mut self, sample_count: usize, ports: &[&'a PortConnection<'a>])

Runs the plugin on a number of samples, given the connected ports.

Provided Methods

fn activate(&mut self)

The plugin instance must reset all state information dependent on the history of the plugin instance here. Will be called before run is called for the first time.

fn deactivate(&mut self)

Indicates the plugin is no longer live.